Actually support what we advertised; new version 0.5
This is a fix for LaTeX errors that would occur when using version 0.4
of the template because the starred versions of the figure commands were
not defined correctly.

Also, this brings in v. 0.5 which includes support for figures around
which text will wrap. This is useful with smaller figures and is known
from text processing software such as MS Word, LibreOffice, etc.

Currently, custom placing parameters for the figures are not supported.