Use \linespread=1 for footnotes
As required by the Austrian Federal Ministry of Education, footnotes in
this document template now use a linespread of 1 as opposed to 1.5. This
was incorrect behaviour before (because the rest of the document is
required to be typeset using one-and-a-half linespread and I simply
forgot to tell KOMA-Script to use single linespread for footnotes).

Also, this commit bumps the version shipped up to 0.7 to prepare for the
next release.