Add a default inclusivity statement for usage of generic masculines
As some subjects require the frequent usage of person-related pronouns
or domain-specific, non-neutral nouns, it can come in handy to use a
construct which is known as "Generisches Maskulinum" (generic masculine)
in German. However, in doing so, the text appears less inclusive in
relation to any other gender.

To counter this, any student is allowed to prepend a "Gendererklärung"
("inclusivity statement" is a very bad translation) to their paper. This
commit adds the command \gendererklaerung (similar to
\selbststaendigkeitserklaerung) which outputs a very generic default
text. This is not yet documented, though.