Add a default spot for \gendererklaerung
To ease the usage of the newly introduced \gendererklaerung command,
this commit adds an example of calling it to the vwa.tex template.

Note that generating an inclusivity statement is generally discouraged
as it implicitly means that using a gender-neutral form of all nouns is
not worth the hassle to the author.